Gather all methods that are related to queries and commits in one file, use a constructor in the __init__ and import a database instance
