+++ src/user_scraper.py
@@ -117,3 +117,32 @@ class UserScraper:
                 tags[index].append(int(tag.text.replace("\n", " ").split()[1].replace(",", "")))
                 tags[index].append(int(tag.text.replace("\n", " ").split()[3].replace(",", "")))
             return tags
+
+
+    @property
+    def user_info(self):
+        """
+        property of user general data for table UserT
+        """
+        return {
+            'rank': self._rank,
+            'name': self._name,
+            'member_since': self._member_since,
+            'profile_views': self._profile_views,
+            'answers': self._answers,
+            'people_reached': self._people_reached,
+        }
+
+    @property
+    def location(self):
+        """
+        property of user location data
+        """
+        return {'country': self._country, 'continent': self._continent}
+
+    @property
+    def reputation(self):
+        """
+        property of user reputation data
+        """
+        return {**self._reputation_hist, 'reputation_now': self._reputation_now}
